requirements.txt
streamlit
python-dotenv
google-generativeai
pinecone
langchain
langchain-google-genai
langchain-community
